Issue Solved #68 (#75)
* frontend for code reviewer

* Code Visibility blurred and Output window added

* merging conflicts
